requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
click==8.1.7
comm==0.1.3
contourpy==1.1.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
exceptiongroup==1.2.1
executing==1.2.0
fastapi==0.110.3
filelock==3.12.2
fonttools==4.42.1
h11==0.14.0
idna==3.7
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
joblib==1.4.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipenv==2023.7.23
platformdirs==3.9.1
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
regex==2024.4.28
scikit-learn==1.4.2
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.37.2
textblob==0.18.0.post0
threadpoolctl==3.5.0
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.11.0
tzdata==2023.3
uvicorn==0.29.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
wcwidth==0.2.6
dask==2024.4.2